Add files via upload
The first assignment I would like to share is my Jelly Monsters math game. This was a group project in which we built a web app and game from the ground up throughout the entire semester. This was a significant assignment for me because while learning new languages and software I was also learning how to build a project with a team, build a product using agile methods, and design using user driven development. 
Our assignment was to create a game for 1st or 2nd graders to learn basic addition and subtraction. We decided to host the game using a web app, and created sign -in and authentication pages, and function for both students and teacher.